+++ src/chains/base.ts
@@ -34,6 +34,9 @@ export enum ChainKey {
   SEI = 'sei',
   FRA = 'fra',
   TAI = 'tai',
+  GRA = 'gra',
+  IMX = 'imx',
+  KAI = 'kai',
 
   // None-EVM
   SOL = 'sol',
@@ -82,6 +85,9 @@ export enum ChainId {
   SEI = 1329,
   FRA = 252,
   TAI = 167000,
+  GRA = 1625,
+  IMX = 13371,
+  KAI = 8217,
 
   // None-EVM (IDs are made up by the LI.FI team)
   SOL = 1151111081099710,