+++ CHANGELOG.md
@@ -2,6 +2,95 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [15.11.0](https://github.com/lifinance/types/compare/v15.10.2...v15.11.0) (2024-09-12)
+
+
+### Features
+
+* added diamond address in /chains ([#278](https://github.com/lifinance/types/issues/278)) ([b7af9bb](https://github.com/lifinance/types/commit/b7af9bb1998081d30084effa978dc82ccdf82fd2))
+
+### [15.10.2](https://github.com/lifinance/types/compare/v15.10.1...v15.10.2) (2024-09-10)
+
+
+### Bug Fixes
+
+* kai chainkey ([#277](https://github.com/lifinance/types/issues/277)) ([7e652b8](https://github.com/lifinance/types/commit/7e652b82e56ea19f33ba0380ecc68d6a7360cb76))
+
+### [15.10.1](https://github.com/lifinance/types/compare/v15.8.0...v15.10.1) (2024-09-10)
+
+
+### Features
+
+* add includedSteps array to full status response ([#275](https://github.com/lifinance/types/issues/275)) ([de01d0c](https://github.com/lifinance/types/commit/de01d0ca58eb7e67c20d4e7e4b3a8240bfeda190))
+* add klaytn chain ([#273](https://github.com/lifinance/types/issues/273)) ([107a3ff](https://github.com/lifinance/types/commit/107a3ffad58e25e81a101af9e441d6e263902632))
+
+## [15.10.0](https://github.com/lifinance/types/compare/v15.8.0...v15.10.0) (2024-09-10)
+
+
+### Features
+
+* add includedSteps array to full status response ([#275](https://github.com/lifinance/types/issues/275)) ([de01d0c](https://github.com/lifinance/types/commit/de01d0ca58eb7e67c20d4e7e4b3a8240bfeda190))
+
+## [15.9.0](https://github.com/lifinance/types/compare/v15.8.0...v15.9.0) (2024-09-05)
+
+
+### Features
+
+* add klaytn chain ([0a76b5c](https://github.com/lifinance/types/commit/0a76b5c5e3aba4076e0dcbca559d9e6e98382d16))
+* from amount in contract calls ([#272](https://github.com/lifinance/types/issues/272)) ([4788192](https://github.com/lifinance/types/commit/47881922b6366b531219c3712375f13cd8c2cfdf))
+
+
+### Bug Fixes
+
+* icon and chain name for Kaia ([a2c60f5](https://github.com/lifinance/types/commit/a2c60f53522f675b41ea3795e0b42c19a0584c47))
+
+## [15.8.0](https://github.com/lifinance/types/compare/v15.7.0...v15.8.0) (2024-09-04)
+
+
+### Features
+
+* add timing parameters ([#271](https://github.com/lifinance/types/issues/271)) ([711deeb](https://github.com/lifinance/types/commit/711deeb12d2e7d9f5c11403c9eb240a895de686e))
+
+## [15.7.0](https://github.com/lifinance/types/compare/v15.6.0...v15.7.0) (2024-09-03)
+
+
+### Features
+
+* add quote toAmount req interface ([#274](https://github.com/lifinance/types/issues/274)) ([68d2a8f](https://github.com/lifinance/types/commit/68d2a8f1499e6b551deee01cbf084fb6b04cd8fe))
+
+## [15.6.0](https://github.com/lifinance/types/compare/v15.4.0...v15.6.0) (2024-09-03)
+
+
+### Features
+
+* from amount in contract calls ([#272](https://github.com/lifinance/types/issues/272)) ([0cd6f86](https://github.com/lifinance/types/commit/0cd6f86ef7949730c635de425da055a93379a5a7))
+
+## [15.5.0](https://github.com/lifinance/types/compare/v15.4.0...v15.5.0) (2024-09-03)
+
+
+### Features
+
+* from amount in contract calls ([#272](https://github.com/lifinance/types/issues/272)) ([0cd6f86](https://github.com/lifinance/types/commit/0cd6f86ef7949730c635de425da055a93379a5a7))
+
+## [15.4.0](https://github.com/lifinance/types/compare/v15.3.2...v15.4.0) (2024-08-29)
+
+
+### Features
+
+* add pol coin ([#270](https://github.com/lifinance/types/issues/270)) ([4ca1504](https://github.com/lifinance/types/commit/4ca1504032ff1e129a2065bf9c7a4735dc23e974))
+
+### [15.3.2](https://github.com/lifinance/types/compare/v15.3.1...v15.3.2) (2024-08-27)
+
+### [15.3.1](https://github.com/lifinance/types/compare/v15.3.0...v15.3.1) (2024-08-27)
+
+## [15.3.0](https://github.com/lifinance/types/compare/v15.1.0...v15.3.0) (2024-08-14)
+
+
+### Features
+
+* add gravity assets ([#264](https://github.com/lifinance/types/issues/264)) ([e9ccbe3](https://github.com/lifinance/types/commit/e9ccbe3f9e4cb1eec7b4b8f94591e54d1f3b1913))
+* add new chain ([#267](https://github.com/lifinance/types/issues/267)) ([5041f55](https://github.com/lifinance/types/commit/5041f55a874c73d1bbf50530752846312c7159fd))
+
 ## [15.1.0](https://github.com/lifinance/types/compare/v15.0.0...v15.1.0) (2024-07-29)